+++ tests/__init__.py
@@ -1,1 +1,0 @@
-"""Various tests for the **Data-CAT** package."""