+++ tests/test_context_managers.py
@@ -1,13 +1,14 @@
 """Tests for :mod:`dataCAT.context_managers`."""
 
 from os.path import join
+from functools import partial
 
 import numpy as np
 import pandas as pd
 
 from assertionlib import assertion
 
-from dataCAT.context_managers import (MetaManager, OpenYaml, OpenLig, OpenQD)
+from dataCAT.context_managers import OpenYaml, OpenLig, OpenQD
 
 PATH = join('tests', 'test_files')
 PATH1 = join(PATH, 'qd.csv')
@@ -16,17 +17,17 @@ PATH1 = join(PATH, 'qd.csv')
 def test_openyaml() -> None:
     """Test :class:`dataCAT.context_managers.OpenYaml`."""
     path = join(PATH, 'settings.yaml')
-    manager = MetaManager(path, OpenYaml)
+    manager = partial(OpenYaml, path)
 
     ref = {None: [None], 'RDKit_2019.03.2': ['UFF'], 'type': [{'ams': {'GeometryOptimization': {'MaxIterations': 500}, 'Task': 'GeometryOptimization'}, 'input': {'ams': {'constraints': {'atom': [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99, 100, 101, 102, 103, 104, 105, 106, 107, 108, 109, 110, 111, 112, 113, 114, 115, 116, 117, 118, 119, 120, 121, 122, 123, 125, 130, 135, 140, 145, 150, 155, 160, 165, 170, 175, 180, 185, 190, 195, 200, 205, 210, 215, 220, 225, 230, 235, 240, 245, 250]}, 'system': {'bondorders': {'_1': None}}}, 'uff': {'library': 'uff'}}}]}  # noqa
-    with manager.open(write=False) as s:
+    with manager(write=False) as s:
         assertion.eq(s, ref)
 
 
 def test_openlig() -> None:
     """Test :class:`dataCAT.context_managers.OpenLig`."""
     path = join(PATH, 'ligand.csv')
-    manager = MetaManager(path, OpenLig)
+    manager = partial(OpenLig, path)
 
     idx = pd.MultiIndex.from_tuples(
         [('-', '-'), ('C[O-]', 'O2'), ('CC[O-]', 'O3'), ('CCC[O-]', 'O4')],
@@ -49,7 +50,7 @@ def test_openlig() -> None:
     ref['opt'] = ref['opt'].astype(bool)
     ref['hdf5 index'] = ref['hdf5 index'].astype(int)
 
-    with manager.open(write=False) as df:
+    with manager(write=False) as df:
         for k in ref:
             np.testing.assert_array_equal(df[k], ref[k])
         np.testing.assert_array_equal(df.index, ref.index)
@@ -59,7 +60,7 @@ def test_openlig() -> None:
 def test_openqd() -> None:
     """Test :class:`dataCAT.context_managers.OpenQD`."""
     path = join(PATH, 'qd.csv')
-    manager = MetaManager(path, OpenQD)
+    manager = partial(OpenQD, path)
 
     idx = pd.MultiIndex.from_tuples(
         [('-', '-', '-', '-'),
@@ -87,7 +88,7 @@ def test_openqd() -> None:
     ref['hdf5 index'] = ref['hdf5 index'].astype(int)
     ref['ligand count'] = ref['ligand count'].astype(int)
 
-    with manager.open(write=False) as df:
+    with manager(write=False) as df:
         for k in ref:
             np.testing.assert_array_equal(df[k], ref[k])
         np.testing.assert_array_equal(df.index, ref.index)