+++ tests/test_database.py
@@ -1,78 +1,43 @@
 """Tests for the :class:`dataCAT.database.Database` class."""
 
-from os.path import join
+from os.path import join, abspath
 import h5py
 import numpy as np
 import pandas as pd
 
 from assertionlib import assertion
 
+from CAT.workflows import MOL, HDF5_INDEX, OPT
 from dataCAT.database import Database
 from dataCAT.context_managers import OpenLig, OpenQD, OpenYaml
 
 PATH = join('tests', 'test_files', 'database')
 DB = Database(PATH)
 
-MOL = ('mol', '')
-HDF5_INDEX = ('hdf5 index', '')
-OPT = ('opt', '')
-
 
 def test_init() -> None:
     """Test :meth:`dataCAT.database.Database.__init__`."""
-    assertion.eq(DB.dirname, PATH)
+    assertion.eq(DB.dirname, abspath(PATH))
 
-    assertion.eq(DB.csv_lig.filename, join(PATH, 'ligand_database.csv'))
-    assertion.is_(DB.csv_lig.manager, OpenLig)
+    assertion.eq(DB.csv_lig.keywords['filename'], abspath(join(PATH, 'ligand_database.csv')))
+    assertion.is_(DB.csv_lig.func, OpenLig)
 
-    assertion.eq(DB.csv_qd.filename, join(PATH, 'qd_database.csv'))
-    assertion.is_(DB.csv_qd.manager, OpenQD)
+    assertion.eq(DB.csv_qd.keywords['filename'], abspath(join(PATH, 'qd_database.csv')))
+    assertion.is_(DB.csv_qd.func, OpenQD)
 
-    assertion.eq(DB.yaml.filename, join(PATH, 'job_settings.yaml'))
-    assertion.is_(DB.yaml.manager, OpenYaml)
+    assertion.eq(DB.yaml.keywords['filename'], abspath(join(PATH, 'job_settings.yaml')))
+    assertion.is_(DB.yaml.func, OpenYaml)
 
-    assertion.eq(DB.hdf5.filename, join(PATH, 'structures.hdf5'))
-    assertion.is_(DB.hdf5.manager, h5py.File)
+    assertion.eq(DB.hdf5.args[0], abspath(join(PATH, 'structures.hdf5')))
+    assertion.is_(DB.hdf5.func, h5py.File)
 
     assertion.is_(DB.mongodb, None)
-
-
-def test_str() -> None:
-    """Test :meth:`dataCAT.database.Database.__str__`."""
-    str_list = str(DB).split('\n')
-
-    assertion.eq(str_list[0], 'Database(')
-    assertion.eq(str_list[1], f'    dirname = {repr(PATH)},')
-    assertion.eq(str_list[-2], '    mongodb = None')
-    assertion.eq(str_list[-1], ')')
-
-    for item in str_list[2:-2]:
-        assertion.contains(item, 'MetaManager(filename=')
-
-
-def test_repr() -> None:
-    """Test :meth:`dataCAT.database.Database.__repr__`."""
-    str_list = repr(DB).split('\n')
-
-    assertion.eq(str_list[0], 'Database(')
-    assertion.eq(str_list[1], f'    dirname = {repr(PATH)},')
-    assertion.eq(str_list[-2], '    mongodb = None')
-    assertion.eq(str_list[-1], ')')
-
-    for item in str_list[2:-2]:
-        assertion.contains(item, 'MetaManager(filename=')
 
 
 def test_eq() -> None:
     """Test :meth:`dataCAT.database.Database.__eq__`."""
     db2 = Database(PATH)
     assertion.eq(db2, DB)
-
-
-def test_contains() -> None:
-    """Test :meth:`dataCAT.database.Database.__contains__`."""
-    assertion.contains(DB, 'mongodb')
-    assertion.contains(DB, 'bob', invert=True)
 
 
 def test_parse_database() -> None: